Allow declarative config of OAuth2 component
Implement a new middleware factory
`loopback-component-oauth2#authenticate` that can be configured
in `server/middleware.json` and which invokes the authentication
middleware created by `loopback-component-oauth2(app)`.

Modify the main exported function to store the list of OAuth2
handlers/endpoints on the app object, so that the new middleware can
retrieve it from there.