This change is the addition of an option called "oauthACLgateway".
If **oauthACLgateway** is *true*, we will check if the *access_token* provided (to query or body) is a user token. If the token is a user token (exists in AccessToken model), we skip oauth authentication.
If no token is provided, we skip oauth authentication. Finally, if the option is omitted or false, the module will work as the original.